Month object new attribute
Month object has a new attribute named doctor_shift.
It is a string array of 31 elements in which will be injected the doctor's name doing the shift.
A new function named "InjectShifts" has been defined, which takes each doctor's manual shift and puts his name in the doctor_shift attribute of the month object.