Credits and weekend calculation + manual entry
The tool now calculates the number of weekends and therefore credits for each doctor, based on rank and days off.
Afterwards, the head doctor can manually modify the value of weekends/credits for each doctor.

The tool will then take those values and insert them into the doctor_list object so that the OR-Tool can enter them as constraints.